chore: initial monorepo
