+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 dist/
 docs/
 *.tgz
+.pnpm-debug.log
