+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zarrita",
+  "name": "@zarrita/monorepo",
   "type": "module",
   "license": "MIT",
   "version": "0.2.2",
@@ -12,7 +12,7 @@
     "test": "uvu -r tsm tests",
     "check": "tsc --noEmit",
     "prepack": "mkdist --declaration",
-    "fmt": "deno fmt --options-use-tabs --options-line-width 90 src tests README.md"
+    "fmt": "deno fmt --options-use-tabs --options-line-width 90 packages/"
   },
   "files": [
     "dist/"
@@ -52,18 +52,14 @@
   },
   "dependencies": {
     "@types/ndarray": "^1.0.11",
-    "@types/wicg-file-system-access": "^2020.9.5",
     "ndarray": "^1.0.19",
     "ndarray-ops": "^1.2.2",
-    "numcodecs": "^0.2.2",
-    "reference-spec-reader": "^0.1.3",
-    "unzipit": "^1.3.6"
+    "numcodecs": "^0.2.2"
   },
   "devDependencies": {
-    "@types/node": "^16.10.2",
     "mkdist": "^0.3.5",
     "tsm": "^2.2.1",
-    "typescript": "^4.6.2",
+    "typescript": "4.7.2",
     "uvu": "^0.5.3"
   }
 }
