+++ packages/core/package.json
@@ -1,0 +1,13 @@
+{
+	"name": "@zarrita/core",
+	"version": "0.0.0",
+	"type": "module",
+	"main": "./src/index.ts",
+	"types": "./src/index.ts",
+	"scripts": {
+		"check": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@zarrita/types": "workspace:*"
+	}
+}
