+++ packages/core/src/custom-arrays.ts
