+++ packages/core/src/errors.ts
