+++ packages/storage/src/index.ts
@@ -1,0 +1,5 @@
+export { default as FetchStore } from "./fetch.js";
+export { default as FileSystemStore } from "./fs.js";
+export { default as MemoryStore } from "./mem.js";
+export { default as ReferenceStore } from "./ref.js";
+export { default as ZipFileStore } from "./zip.js";
