+++ packages/storage/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.json",
+	"include": ["src/**/*"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"types": ["@types/wicg-file-system-access", "@types/node"]
+	}
+}
