+++ packages/v2/package.json
@@ -1,0 +1,12 @@
+{
+	"name": "@zarrita/v2",
+	"main": "src/index.ts",
+	"types": "src/index.ts",
+	"scripts": {
+		"check": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@zarrita/core": "workspace:*",
+		"@zarrita/types": "workspace:*"
+	}
+}
