+++ packages/v3/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"include": ["src/**/*"],
+	"compilerOptions": {
+		"outDir": "./dist"
+	}
+}
