--- src/index.ts
@@ -1,7 +1,0 @@
-export * as v2 from "./v2";
-export * as v3 from "./v3";
-export * as ops from "./ops";
-export * as ndarray from "./ndarray";
-export { slice } from "./lib/util";
-export { registry } from "./lib/codec-registry";
-export * from "./types";
