--- src/storage/index.ts
@@ -1,5 +1,0 @@
-export { default as FetchStore } from "./fetch";
-export { default as FileSystemStore } from "./fs";
-export { default as MemoryStore } from "./mem";
-export { default as ReferenceStore } from "./ref";
-export { default as ZipFileStore } from "./zip";
